GOATS-329: Refactor codebase and clean up directories.
