+++ .github/workflows/build_conda_env.yml
@@ -26,7 +26,6 @@ jobs:
         with:
           mamba-version: "*"
           python-version: ${{ matrix.python-version }}
-          miniforge-variant: Mambaforge
           miniforge-version: latest
           activate-environment: goats-env
           use-mamba: true
