+++ doc/changes/GOATS-329.change.md
@@ -1,0 +1,1 @@
+Refactored codebase for better organization.
