+++ src/goats_tom/tasks/__init__.py
@@ -1,3 +1,4 @@
-from .tasks import download_goa_files, run_dragons_reduce
+from .download_goa_files import download_goa_files
+from .run_dragons_reduce import run_dragons_reduce
 
 __all__ = ["download_goa_files", "run_dragons_reduce"]
