+++ src/goats_tom/views/__init__.py
@@ -1,5 +1,10 @@
 from .base_recipe import BaseRecipeViewSet
+from .brokerquery_name import update_brokerquery_name
+from .dataproduct_delete import DataProductDeleteView
 from .dataproducts import DataProductsViewSet
+from .delete_observation_dataproducts import DeleteObservationDataProductsView
+from .downloads import recent_downloads
+from .dragons import DRAGONSView
 from .dragons_caldb import DRAGONSCaldbViewSet
 from .dragons_data import DRAGONSDataViewSet
 from .dragons_files import DRAGONSFilesViewSet
@@ -7,27 +12,16 @@ from .dragons_output_files import DRAGONSOutputFilesViewSet
 from .dragons_recipes import DRAGONSRecipesViewSet
 from .dragons_reduce import DRAGONSReduceViewSet
 from .dragons_runs import DRAGONSRunsViewSet
+from .goa_login import GOALoginView
+from .goa_query_form import GOAQueryFormView
+from .keys import ManageKeysView, activate_user_key, create_key, delete_key
+from .observation_record_delete import ObservationRecordDeleteView
+from .observation_record_detail import ObservationRecordDetailView
+from .receive_query import receive_query
 from .recipes_module import RecipesModuleViewSet
-from .views import (
-    DeleteObservationDataProductsView,
-    DRAGONSView,
-    GOALoginView,
-    GOAQueryFormView,
-    GOATSDataProductDeleteView,
-    GOATSObservationRecordDetailView,
-    GOATSTargetDeleteView,
-    ManageKeysView,
-    ObservationRecord,
-    ObservationRecordDeleteView,
-    UserGenerateTokenView,
-    activate_user_key,
-    create_key,
-    delete_key,
-    ongoing_tasks,
-    receive_query,
-    recent_downloads,
-    update_brokerquery_name,
-)
+from .target_delete import TargetDeleteView
+from .tasks import ongoing_tasks
+from .user_generate_token import UserGenerateTokenView
 
 __all__ = [
     "DRAGONSRecipesViewSet",
@@ -37,17 +31,18 @@ __all__ = [
     "DeleteObservationDataProductsView",
     "GOALoginView",
     "GOAQueryFormView",
-    "GOATSDataProductDeleteView",
-    "GOATSObservationRecordDetailView",
-    "GOATSTargetDeleteView",
+    "DataProductDeleteView",
+    "ObservationRecordDetailView",
+    "TargetDeleteView",
     "ManageKeysView",
-    "ObservationRecord",
     "UserGenerateTokenView",
     "activate_user_key",
     "create_key",
     "delete_key",
-    "ongoing_tasks",
+    "tasks",
     "receive_query",
+    "downloads",
+    "ongoing_tasks",
     "recent_downloads",
     "update_brokerquery_name",
     "ObservationRecordDeleteView",
